django_api_project/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
celery
redis
django-celery-results
python-dotenv
gunicorn
whitenoise
drf-yasg
django-filter
